refactor: Move ctx statedb to separate package, rename hybridstatedb to statedb
