+++ x/evm/ctxstatedb/access_list.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package hybridstatedb
+package ctxstatedb
 
 import (
 	"github.com/ethereum/go-ethereum/common"
