+++ x/evm/ctxstatedb/store_test.go
@@ -1,11 +1,11 @@
-package statedb_test
+package ctxstatedb_test
 
 import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/evmos/ethermint/x/evm/statedb"
+	statedb "github.com/evmos/ethermint/x/evm/ctxstatedb"
 	"github.com/stretchr/testify/require"
 )
