+++ x/evm/statedb/interfaces.go
@@ -37,6 +37,8 @@ type ExtStateDB interface {
 type Keeper interface {
 	// Read methods
 	GetAccount(ctx sdk.Context, addr common.Address) *Account
+	GetBalance(ctx sdk.Context, addr common.Address) *big.Int
+	GetAccountNumber(ctx sdk.Context, addr common.Address) (uint64, bool)
 	GetState(ctx sdk.Context, addr common.Address, key common.Hash) common.Hash
 	GetCode(ctx sdk.Context, codeHash common.Hash) []byte
 	// the callback returns false to break early
@@ -44,11 +46,7 @@ type Keeper interface {
 
 	// Write methods, only called by `StateDB.Commit()`
 	SetAccount(ctx sdk.Context, addr common.Address, account Account) error
-	SetState(ctx sdk.Context, addr common.Address, key, value common.Hash)
-	UnsetState(ctx sdk.Context, addr common.Address, key common.Hash) error
-
-	ReassignAccountNumbers(ctx sdk.Context, addrs []common.Address) error
-
+	SetState(ctx sdk.Context, addr common.Address, key common.Hash, value common.Hash)
 	SetCode(ctx sdk.Context, codeHash []byte, code []byte)
 	SetBalance(ctx sdk.Context, addr common.Address, amount *big.Int) error
 	DeleteAccount(ctx sdk.Context, addr common.Address) error
