+++ x/evm/statedb/state_object.go
@@ -17,6 +17,7 @@ package statedb
 
 import (
 	"bytes"
+	"math/big"
 	"sort"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -24,6 +25,28 @@ import (
 )
 
 var emptyCodeHash = crypto.Keccak256(nil)
+
+// Account is the Ethereum consensus representation of accounts.
+// These objects are stored in the storage of auth module.
+type Account struct {
+	// Balance is *not* included as it is managed by bank
+	Nonce    uint64
+	CodeHash []byte
+
+	AccountNumber uint64
+}
+
+// NewEmptyAccount returns an empty account.
+func NewEmptyAccount() *Account {
+	return &Account{
+		CodeHash: emptyCodeHash,
+	}
+}
+
+// IsContract returns if the account contains contract code.
+func (acct Account) IsContract() bool {
+	return !bytes.Equal(acct.CodeHash, emptyCodeHash)
+}
 
 // Storage represents in-memory cache/buffer of contract storage.
 type Storage map[common.Hash]common.Hash
@@ -40,4 +63,209 @@ func (s Storage) SortedKeys() []common.Hash {
 		return bytes.Compare(keys[i].Bytes(), keys[j].Bytes()) < 0
 	})
 	return keys
+}
+
+// stateObject is the state of an acount
+type stateObject struct {
+	db *StateDB
+
+	account Account
+	code    []byte
+
+	// state storage
+	originStorage Storage
+	dirtyStorage  Storage
+
+	address common.Address
+
+	// flags
+	dirtyCode    bool
+	suicided     bool
+	isNew        bool // created in this tx
+	dirtyBalance bool // balance has changed
+}
+
+// newObject creates a state object.
+func newObject(
+	db *StateDB,
+	address common.Address,
+	account Account,
+	isNew bool,
+) *stateObject {
+	if account.CodeHash == nil {
+		account.CodeHash = emptyCodeHash
+	}
+	return &stateObject{
+		db:            db,
+		address:       address,
+		account:       account,
+		originStorage: make(Storage),
+		dirtyStorage:  make(Storage),
+		isNew:         isNew,
+		dirtyBalance:  false,
+	}
+}
+
+// empty returns whether the account is considered empty.
+func (s *stateObject) empty() bool {
+	return s.account.Nonce == 0 &&
+		s.Balance().Sign() == 0 &&
+		bytes.Equal(s.account.CodeHash, emptyCodeHash)
+}
+
+func (s *stateObject) markSuicided() {
+	s.suicided = true
+}
+
+// AddBalance adds amount to s's balance.
+// It is used to add funds to the destination account of a transfer.
+func (s *stateObject) AddBalance(amount *big.Int) {
+	if amount.Sign() == 0 {
+		return
+	}
+	s.SetBalance(new(big.Int).Add(s.Balance(), amount))
+}
+
+// SubBalance removes amount from s's balance.
+// It is used to remove funds from the origin account of a transfer.
+func (s *stateObject) SubBalance(amount *big.Int) {
+	if amount.Sign() == 0 {
+		return
+	}
+	s.SetBalance(new(big.Int).Sub(s.Balance(), amount))
+}
+
+// SetBalance update account balance.
+func (s *stateObject) SetBalance(amount *big.Int) {
+	// s.db.journal.append(balanceChange{
+	// 	account: &s.address,
+	// 	prev:    new(big.Int).Set(s.account.Balance),
+	// })
+	s.setBalance(amount)
+}
+
+func (s *stateObject) setBalance(amount *big.Int) {
+	if err := s.db.keeper.SetBalance(s.db.ctx.CurrentCtx(), s.address, amount); err != nil {
+		// TODO: Move to Commit()
+		panic(err)
+	}
+
+	// SetBalance will create the account if it doesn't exist yet, so we need to
+	// mark it as dirty to keep track of which accounts were created outside of
+	// Commit()
+	s.dirtyBalance = true
+}
+
+//
+// Attribute accessors
+//
+
+// Returns the address of the contract/account
+func (s *stateObject) Address() common.Address {
+	return s.address
+}
+
+// Code returns the contract code associated with this object, if any.
+func (s *stateObject) Code() []byte {
+	if s.code != nil {
+		return s.code
+	}
+	if bytes.Equal(s.CodeHash(), emptyCodeHash) {
+		return nil
+	}
+	code := s.db.keeper.GetCode(s.db.ctx.InitialCtx(), common.BytesToHash(s.CodeHash()))
+	s.code = code
+	return code
+}
+
+// CodeSize returns the size of the contract code associated with this object,
+// or zero if none.
+func (s *stateObject) CodeSize() int {
+	return len(s.Code())
+}
+
+// SetCode set contract code to account
+func (s *stateObject) SetCode(codeHash common.Hash, code []byte) {
+	prevcode := s.Code()
+	s.db.journal.append(codeChange{
+		account:  &s.address,
+		prevhash: s.CodeHash(),
+		prevcode: prevcode,
+	})
+	s.setCode(codeHash, code)
+}
+
+func (s *stateObject) setCode(codeHash common.Hash, code []byte) {
+	s.code = code
+	s.account.CodeHash = codeHash[:]
+	s.dirtyCode = true
+}
+
+// SetCode set nonce to account
+func (s *stateObject) SetNonce(nonce uint64) {
+	s.db.journal.append(nonceChange{
+		account: &s.address,
+		prev:    s.account.Nonce,
+	})
+	s.setNonce(nonce)
+}
+
+func (s *stateObject) setNonce(nonce uint64) {
+	s.account.Nonce = nonce
+}
+
+// CodeHash returns the code hash of account
+func (s *stateObject) CodeHash() []byte {
+	return s.account.CodeHash
+}
+
+// Balance returns the balance of account
+func (s *stateObject) Balance() *big.Int {
+	// Balance tracking uses the current ctx state, NOT journal state as
+	// precompiles can modify balances directly.
+	return s.db.keeper.GetBalance(s.db.ctx.CurrentCtx(), s.address)
+}
+
+// Nonce returns the nonce of account
+func (s *stateObject) Nonce() uint64 {
+	return s.account.Nonce
+}
+
+// GetCommittedState query the committed state
+func (s *stateObject) GetCommittedState(key common.Hash) common.Hash {
+	if value, cached := s.originStorage[key]; cached {
+		return value
+	}
+	// If no live objects are available, load it from keeper
+	value := s.db.keeper.GetState(s.db.ctx.InitialCtx(), s.Address(), key)
+	s.originStorage[key] = value
+	return value
+}
+
+// GetState query the current state (including dirty state)
+func (s *stateObject) GetState(key common.Hash) common.Hash {
+	if value, dirty := s.dirtyStorage[key]; dirty {
+		return value
+	}
+	return s.GetCommittedState(key)
+}
+
+// SetState sets the contract state
+func (s *stateObject) SetState(key common.Hash, value common.Hash) {
+	// If the new value is the same as old, don't set
+	prev := s.GetState(key)
+	if prev == value {
+		return
+	}
+	// New value is different, update and journal the change
+	s.db.journal.append(storageChange{
+		account:  &s.address,
+		key:      key,
+		prevalue: prev,
+	})
+	s.setState(key, value)
+}
+
+func (s *stateObject) setState(key, value common.Hash) {
+	s.dirtyStorage[key] = value
 }
